Avoid putting common code in package __init__.
